+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "youtube.get-video-info",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "A parser for the get_video_info request on Youtube (no API required)",
-  "main": "youtube.get-video.info.js",
+  "main": "youtube.get-video-info.js",
   "scripts": {
     "test": "node test.js"
   },
